requirements.txt
liboqs-python
qiskit
qiskit-ibm-runtime
numpy
openai
flask
flask-cors
requests
gunicorn
google-generativeai
